Harden plugin ABI and workflow runtime boundaries
- bump bundled plugin manifests to ABI version 4
- strengthen plugin TUI surface handling and workflow execution limits
- improve workflow persistence and server session-search integration

+++ plugins/workflow-plugin/src/lib.rs
@@ -1041,8 +1041,7 @@ fn reference_template_run_limits(
         .ok_or_else(|| "reference template iteration_limit is required".to_string())?;
     limits.cycle_cap = u32::try_from(iteration_limit)
         .map_err(|_| "reference template iteration_limit exceeds u32".to_string())?;
-    limits.node_execution_cap = limits
-        .cycle_cap
+    limits.node_execution_cap = u64::from(limits.cycle_cap)
         .checked_mul(16)
         .ok_or_else(|| "reference template node execution limit overflow".to_string())?;
     Ok(limits)
